+++ src/core/gzip.h
@@ -1,6 +1,6 @@
 #ifndef __gzip_h
 #define __gzip_h
 
-unsigned long int decompress_data(unsigned char* abSrc, int nLenSrc, unsigned char** abDst, int nLenDst);
+unsigned long int decompress_data(unsigned char* abSrc, unsigned long nLenSrc, unsigned char** abDst, int nLenDst);
   
 #endif
