+++ src/core/stepper.c
@@ -17,7 +17,7 @@ error:
 }
 
 void Stepper_set_steps_per_second(Stepper *stepper,
-        unsigned int steps_per_second) {
+        float steps_per_second) {
     stepper->steps_per_second = steps_per_second;
 
     if (steps_per_second > 0)
@@ -26,15 +26,15 @@ void Stepper_set_steps_per_second(Stepper *stepper,
         stepper->step_skip = 0;
 }
 
-int Stepper_update(Stepper *stepper, long unsigned int step_ticks) {
+unsigned long Stepper_update(Stepper *stepper, long unsigned int step_ticks) {
     if(stepper->step_skip == 0) return 0;
 
-    int accumulation = step_ticks % stepper->step_skip;
-    int pure_steps = step_ticks / stepper->step_skip;
+    unsigned long accumulation = step_ticks % stepper->step_skip;
+    unsigned long pure_steps = step_ticks / stepper->step_skip;
     stepper->stack += pure_steps;
 
     stepper->accumulator += accumulation;
-    int accumulated_steps = stepper->accumulator / stepper->step_skip;
+    unsigned long accumulated_steps = stepper->accumulator / stepper->step_skip;
     stepper->stack += accumulated_steps;
     stepper->accumulator %= stepper->step_skip;
     return pure_steps + accumulated_steps;
@@ -52,8 +52,8 @@ int Stepper_pop(Stepper *stepper) {
     return 1;
 }
 
-int Stepper_clear(Stepper *stepper) {
-    int stack = stepper->stack;
+unsigned long Stepper_clear(Stepper *stepper) {
+    unsigned long stack = stepper->stack;
     stepper->stack = 0;
     return stack;
 }
